# docs/openapi/aether-v0.yaml
openapi: 3.1.0
info:
  title: Aether /api/v0 API
  version: 0.1.0
  contact:
    name: Andres Bott
    url: https://github.com/andresbott/aether
  description: >
    Internal, admin-only server-management API — collections/libraries,
    scanning, tasks/scheduling, and other administration concerns. This is
    NOT the music-client API: browsing, playback, playlists, album/artist
    navigation and search live exclusively under `/rest`, which must stay
    OpenSubsonic-compliant and has no OpenAPI document of its own (a Subsonic
    client does not consume one).


    **Scope.** This document describes the full `/api/v0` surface as mounted
    in `app/router/api_v0.go`: the metadata editor (pictures, raw tags,
    folder/track browsing, identify and identify-album, artist images),
    artists, users, tasks, tokens, radiobrowser, auth, libraries, and
    health/version/me.
servers:
  - url: /api/v0
    description: >
      Every path below is mount-relative. The base is modelled exclusively
      through this `servers` entry, so bumping the version prefix — this is
      `/api/v0` while the API is pre-1.0 — stays a one-line change, with no
      path or operationId edits.
...
security:
  - sessionCookie: []
paths:
...
  /metadata/folders:
    get:
      operationId: listMetadataFolders
      summary: Immediate subfolders of a library directory (the editor's folder-tree picker).
      description: >
        Confined lexically to the library root (no symlink traversal), unlike
        the library path picker elsewhere in the app. A non-blank `q`
        switches the endpoint from a one-level listing to a recursive
        name search of the whole subtree under `path`, capped at 500
        matches; `q` and the response's `truncated` field are both
        meaningful only in that search branch — a plain listing (no `q`)
        carries neither.
      tags: [metadata-browse]
      parameters:
        - $ref: '#/components/parameters/LibraryId'
        - $ref: '#/components/parameters/LibraryPath'
        - name: q
          in: query
          required: false
          description: >
            Folder-name search filter. Blank/omitted (the default) lists
            only the immediate subfolders of `path`; non-blank recursively
            searches the whole subtree under `path` for folders whose name
            matches (see `FolderList.truncated`).
          schema: { type: string }
      responses:
        '200':
          description: >
            Immediate subdirectories, sorted by name — or, when `q` is set,
            every matching folder in the subtree (see `FolderList.truncated`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FolderList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /metadata/artist-folder:
    get:
      operationId: getArtistFolder
      summary: Whether the selected folder is an artist folder, and its current portrait image.
      description: >
        A pure filesystem+tags question, independent of the library index
        and any track selection: eligible when the folder — or the nearest
        ancestor of the selected path — has sub-folders whose tracks are
        tagged with an album artist matching the folder's own name
        (`metadataedit.IsArtistFolder`, the same rule the scanner uses for
        artist-image detection).
      tags: [metadata-artist-image]
      parameters:
        - $ref: '#/components/parameters/LibraryId'
        - $ref: '#/components/parameters/LibraryPath'
      responses:
        '200':
          description: >
            `{eligible:false}` alone when the selected path is not (and has
            no ancestor that is) an artist folder. See `ArtistFolder` for
            the eligible shape.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistFolder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /metadata/artist-image:
    get:
      operationId: getArtistImage
      summary: Bytes of the selected artist folder's current portrait image.
      description: >
        Serves the on-disk file `artistimage.BestInDir` picks (highest
        priority: `artist.*`, then `artistthumb.*`, then `folder.*`)
        directly, rather than any DB-resolved cover — the editor previews
        exactly what it manages. `Cache-Control: no-cache`, so the editor's
        cache-busted `<img>` reload after a change always fetches fresh
        bytes.
      tags: [metadata-artist-image]
      parameters:
        - $ref: '#/components/parameters/LibraryId'
        - $ref: '#/components/parameters/LibraryPath'
      responses:
        '200':
          description: >
            The image's bytes, exactly as stored on disk (no re-encoding).
            Content-Type is derived from the file's extension: normally
            image/jpeg or image/png (the only formats this API's own writes
            ever produce), but any pre-existing artist/artistthumb/folder
            file the editor did not write itself may also be image/gif,
            image/bmp or image/webp — the full set `artistimage`'s folder
            detection recognises.
          headers:
            Cache-Control:
              schema: { type: string, example: "no-cache" }
          content:
            image/*: { schema: { type: string, format: binary } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: >
            The library does not exist, or the selected folder holds no
            artist image. The latter case answers with Go's plain
            `http.NotFound`; like `getPictureImage`'s stale-link 404, the
            router-level error envelope (`app/router/errors.go`) still
            wraps it into the same `application/problem+json` Problem shape
            every other response on this surface uses — only its `detail`
            text differs ("404 page not found" rather than a
            handler-authored message).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    post:
      operationId: setArtistImage
      summary: Write an artist portrait into the selected folder, from an upload or an online pick.
      description: >
        Writes `artist.<ext>` (jpg or png) into the SELECTED folder —
        removing any other-format `artist.*` sibling — from either an
        uploaded file or an online pick downloaded through the
        SSRF-guarded provider chain (see `getArtistImageCandidateInfo`).
        Nothing is written to the library index directly: one
        representative track under the folder is rescanned synchronously so
        the scanner's reconcile re-detects the artist's image (`rescan`
        reports the outcome; a rescan failure is never a write failure —
        the file is already on disk).
      tags: [metadata-artist-image]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/SetArtistImageForm' }
      responses:
        '200':
          description: Written.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplyArtistImageResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '502': { $ref: '#/components/responses/UpstreamError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
    delete:
      operationId: deleteArtistImage
      summary: Remove the selected artist folder's current portrait image.
      description: >
        Removes the same file `getArtistImage` serves, then re-indexes one
        representative track so the scanner's reconcile clears (or
        re-detects a fallback) artist image. `MutationResult` here only
        ever sets `ok`/`rescan` — `slot`/`type` are absent entirely,
        meaningless for an artist-folder image.
      tags: [metadata-artist-image]
      parameters:
        - $ref: '#/components/parameters/LibraryId'
        - $ref: '#/components/parameters/LibraryPath'
      responses:
        '200':
          description: Removed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: >
            The library does not exist, or the selected folder holds no
            artist image to remove (Go's plain `http.NotFound`, wrapped the
            same way as `getArtistImage`'s 404 — see its description).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /libraries/{id}:
    get:
      operationId: getLibrary
      summary: Fetch one library.
      description: >
        Admin. `track_count` is computed fresh on every read (a live query
        against the tracks table), not a cached column.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The library.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Library' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      operationId: updateLibrary
      summary: Replace a library's editable fields.
      description: >
        Admin. `name`/`path`/`exclude_patterns`/`default_view`/`icon`/
        `cover_style` are always fully replaced by the submitted value
        (empty/omitted `default_view`, `icon` and `cover_style` fall back
        the same way createLibrary's do) — this is a full replace, not a
        partial update. `follow_symlinks`/`show_artists` are the only true
        tri-state fields here: omitted (or explicit null) keeps the
        library's current stored value instead of resetting it — see
        LibraryUpdateRequest.

        Changing `path` wipes and re-scans the library's tracks; the
        response's `path_changed` reports whether that happened, and is
        present only then (omitempty on the handler's side) — never
        present-and-false.

        Refused with 409 `config_managed` for a config-provisioned library
        — edit the Libraries section of config.yaml and restart instead.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LibraryUpdateRequest' }
      responses:
        '200':
          description: >
            Updated. The Library shape plus `path_changed` — see the
            operation description.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Library'
                  - type: object
                    properties:
                      path_changed:
                        type: boolean
                        description: >
                          True when this update changed `path`. Omitted
                          entirely — not present-and-false — when the path
                          was unchanged.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      operationId: deleteLibrary
      summary: Delete a library and its tracks.
      description: >
        Admin. Refused with 409 `config_managed` for a config-provisioned
        library, for the same reason as updateLibrary.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Deleted. No body.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /artists/{id}/mbid:
    get:
      operationId: getArtistMBID
      summary: An artist's current MusicBrainz association.
      description: >
        Admin. Reads the artist's stored MusicBrainz association; see
        `setArtistMBID` to change it. `mbArtistId` is empty, not omitted,
        when the artist has no association.
      tags: [artists]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The artist's MusicBrainz id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistMbid' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      operationId: setArtistMBID
      summary: Set (or clear) an artist's MusicBrainz association.
      description: >
        Admin. `mbid` empty clears the association and skips the image
        fetch entirely; a non-empty value must match the MusicBrainz
        identifier format (a UUID), or this answers 400. Setting a
        non-empty `mbid` also triggers a one-shot auto-fetch of the
        artist's image — its outcome is reported in the response's
        `imageFetched`/`fetchError`, never as a request failure: the mbid
        association is saved either way, including when artist-image
        fetching has no provider configured. See `SetArtistMbidResult`.
      tags: [artists]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetArtistMbidRequest' }
      responses:
        '200':
          description: Saved. See `SetArtistMbidResult` for the auto-fetch outcome.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetArtistMbidResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: _aether_login
      description: >
        Native-mode session cookie (see docs/agents/authentication.md).
        `/api/v0` defaults every route to admin-only, so a valid, enabled
        admin session is required for the whole metadata surface. This is the
        only credential `/api/v0` actually validates: no route on this
        surface reads an `Authorization` header, and Aether's PATs
        (bearer-shaped credentials) are verified exclusively on `/rest` (via
        `apiKey`, or `u`+`t`+`s` / `u`+`p`), never here.


        In **proxy-header mode**, this cookie plays no role either — identity
        instead comes from reverse-proxy-injected headers (`Remote-User` /
        `Remote-Groups`), validated by `headerGuard`. That trust is
        established between the proxy and Aether, not chosen by the API
        caller, so it has no client-facing credential to model as an OpenAPI
        security scheme; it is intentionally left undeclared here rather than
        represented by a scheme nothing in the protocol actually sends.
  parameters:
    LibraryId:
      name: library_id
      in: query
      required: true
      description: The collection/library id to resolve paths against.
      schema: { type: integer, format: int64, minimum: 0 }
    LibraryPath:
      name: path
      in: query
      required: false
      description: >
        A directory, library-relative; empty/omitted defaults to the library
        root. Resolved lexically against the library root — an absolute path,
        or one that escapes the root, answers 400.
      schema: { type: string }
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details. `type` is a stable, never-fetched URI whose
        last path segment mirrors the old ad-hoc "code" string.
      properties:
        type:
          type: string
          format: uri
          description: Never dereferenced; only the last path segment is meant to be read.
          example: "https://aether.local/probs/validation_error"
        title:
          type: string
          example: "Validation error"
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: The request path.
      required: [type, title, status]
...
    FolderList:
      type: object
      properties:
        folders:
          type: array
          items: { $ref: '#/components/schemas/Folder' }
        truncated:
          type: boolean
          description: >
            Present only when the request carried a non-blank `q`: `true`
            when the search hit the server's 500-match cap and more folders
            would have matched. Absent (not merely `false`) from a plain,
            non-search listing.
      required: [folders]
...
    ArtistFolder:
      type: object
      description: >
        Whether the SELECTED folder is an artist folder, and the portrait
        image it currently holds. `artist`/`path`/`current_image`/
        `current_image_meta` are omitted — not merely empty — when
        `eligible` is `false`. `current_image` (and, with it,
        `current_image_meta`) is independently omitted whenever the folder
        holds no artist image, even when `eligible` is `true`.
      properties:
        eligible: { type: boolean }
        artist:
          type: string
          description: >
            The artist name (the resolved folder's own basename). Present
            only when eligible.
        path:
          type: string
          description: >
            Library-relative path of the resolved artist folder — the
            selected folder itself, or the nearest ancestor that
            qualifies. Present only when eligible.
        current_image:
          type: string
          description: >
            Filename of the current artist image, e.g. "artist.jpg".
            Present only when one exists.
        current_image_meta:
          allOf:
            - $ref: '#/components/schemas/ImageMeta'
          description: Present only alongside `current_image`.
      required: [eligible]
...
    ApplyArtistImageResult:
      type: object
      properties:
        ok: { type: boolean }
        path:
          type: string
          description: Library-relative path of the written file, e.g. "Radiohead/artist.jpg".
        rescan: { $ref: '#/components/schemas/RescanStatus' }
      required: [ok, path]
...
  responses:
    BadRequest:
      description: Malformed request (bad JSON/multipart, missing/invalid scalar).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    TooManyRequests:
      description: A third-party upstream service is rate-limiting Aether's requests.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UpstreamError:
      description: >
        A third-party upstream service call failed. The same upstream
        classification can also surface as 504 on a timeout (still this
        Problem shape) — see `GatewayTimeout`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    GatewayTimeout:
      description: A third-party upstream service timed out.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: No valid session (native mode) — the caller must authenticate.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted (admin required, or the user is disabled).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    InternalServerError:
      description: >
        An unexpected server-side failure — a store or internal error that is
        not a not-found (e.g. a failed database read/write, a filesystem write
        that did not complete). The generic Problem shape; see `type`/`detail`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }